update player scraping

+++ app/Http/Controllers/ScrapeController.php
@@ -15,14 +15,14 @@ class ScrapeController extends Controller
 {
     public function xflPlayers() {
         $urls = [
-            "Dallas" => "https://www.xfl.com/en-US/articles/dallas-renegades-roster",
+            "Dallas" => "https://www.xfl.com/en-US/teams/dallas/renegades-articles/dallas-renegades-roster",
             "DC" => "https://www.xfl.com/en-US/teams/washington-dc/defenders-articles/dc-defenders-roster",
-            "Houston" => "https://www.xfl.com/en-US/articles/houston-roughnecks-roster",
-            "LA" => "https://www.xfl.com/en-US/articles/la-wildcats-roster",
-            "New York" => "https://www.xfl.com/en-US/articles/new-york-guardians-roster",
-            "St Louis" => "https://www.xfl.com/en-US/articles/st-louis-battlehawks-roster",
-            "Seattle" => "https://www.xfl.com/en-US/articles/seattle-dragons-roster",
-            "Tampa Bay" => "https://www.xfl.com/en-US/articles/tampa-bay-vipers-roster"
+            "Houston" => "https://www.xfl.com/en-US/teams/houston/roughnecks-articles/houston-roughnecks-roster",
+            "LA" => "https://www.xfl.com/en-US/teams/los-angeles/wildcats-articles/la-wildcats-roster",
+            "New York" => "https://www.xfl.com/en-US/teams/new-york/guardians-articles/new-york-guardians-roster",
+            "St Louis" => "https://www.xfl.com/en-US/teams/st-louis/battlehawks-articles/st-louis-battlehawks-roster",
+            "Seattle" => "https://www.xfl.com/en-US/teams/seattle/dragons-articles/seattle-dragons-roster",
+            "Tampa Bay" => "https://www.xfl.com/en-US/teams/tampa-bay/vipers-articles/tampa-bay-vipers-roster"
         ];
 
         foreach ($urls as $team=>$endpoint) {
@@ -39,11 +39,24 @@ class ScrapeController extends Controller
                     $player_college = '';
                     
                     $college_cell_key = 5;
-                    if ($team == "DC") $college_cell_key = 3;
+                    if ($team == "DC") $college_cell_key = 6;
+                    if ($team == "Dallas") $college_cell_key = 6;
+                    if ($team == "LA") $college_cell_key = 6;
                     foreach($element->find('td') as $cellkey=>$cell) {
                         if ($cellkey == 1) {
-                            $player_name = str_replace("&nbsp;","",trim(strip_tags($cell->innertext)));
+                            if ($team == "Dallas" || $team == "DC" || $team == "LA") {
+                                $player_first_name = str_replace("&nbsp;","",trim(strip_tags($cell->innertext)));
+                            } else {
+                                $player_name = str_replace("&nbsp;","",trim(strip_tags($cell->innertext)));
+                            }
                         } else if ($cellkey == 2) {
+                            if ($team == "Dallas" || $team == "DC" || $team == "LA") {
+                                $player_last_name = str_replace("&nbsp;","",trim(strip_tags($cell->innertext)));
+                                $player_name = $player_last_name.", ".$player_first_name;
+                            } else {
+                                $player_position = trim(strip_tags($cell->innertext));
+                            }
+                        } else if ($team == ("Dallas" || "DC" || $team == "LA") && $cellkey == 3) {
                             $player_position = trim(strip_tags($cell->innertext));
                         } else if ($cellkey == $college_cell_key) {
                             $player_college = trim(strip_tags($cell->innertext));
@@ -72,14 +85,15 @@ class ScrapeController extends Controller
         // Deal with comma dupes
         $search = Player::get();
         foreach($search as $player) {
-            //$name_parts = explode(" ", $comma_player->name);
             if (!strpos($player->name, ",") && $player->position != "DEF") {
                 $name_parts = explode(" ",$player->name);
-                $site_name = $name_parts[1].", ".$name_parts[0];
-                $delete = Player::where('name',$site_name)->delete();
-                $update = Player::where('name',$player->name)->update([
-                    'name'=>$site_name
-                ]);
+                if (isset($name_parts[1])) {
+                    $site_name = $name_parts[1].", ".$name_parts[0];
+                    $delete = Player::where('name',$site_name)->delete();
+                    $update = Player::where('name',$player->name)->update([
+                        'name'=>$site_name
+                    ]);
+                }
             }
         }
